revert(rccl): Unit tests for src/dev_runtime.cc (#10981)
