--- projects/rccl/test/test_categories_micro_devruntime.yaml
@@ -1,64 +1,0 @@
-# rccl-UnitTestsMicroDevRuntime — shared/ctest schema.
-# Host-only micro-tests for src/dev_runtime.cc (no GPU, no librccl/HIP runtime); build in Release and Debug.
-# RCCL does not support Windows: exclude_windows matches all tests so generated suites are empty on Windows.
-#
-# Patterns are prefix globs rather than an enumeration of every suite, so adding
-# a suite in an existing family (Devr*, Sym*, Win*, ...) does not silently fall
-# out of CI gating.
-#
-# Tiers:
-#   quick — canonical anchor.
-#   standard — merge quick; override labels.
-#   comprehensive — merge quick; override labels.
-#   full — merge quick; override labels.
-
-test_categories:
-  quick: &micro_devruntime_quick
-    description: "Smoke — host-only src/dev_runtime.cc micro-tests (symmetric memory, windows, devcomm), sub-second."
-    test_patterns:
-      - "AllocAndPopulateSegmentWindowsTest.*"
-      - "Comm*"
-      - "ComputeLsaSizeTest.*"
-      - "DeepCopyDevCommRequirementsTest.*"
-      - "Dev*"
-      - "GinResourcesRequestedTest.*"
-      - "NcclVersionCompatTest.*"
-      - "Sym*"
-      - "Win*"
-    exclude_windows:
-      - "*"
-    labels:
-      - "rccl"
-      - "quick"
-      - "smoke"
-      - "precheck"
-      - "microtest"
-  standard:
-    <<: *micro_devruntime_quick
-    labels:
-      - "rccl"
-      - "standard"
-      - "regression"
-      - "microtest"
-  comprehensive:
-    <<: *micro_devruntime_quick
-    labels:
-      - "rccl"
-      - "comprehensive"
-      - "comprehensive_regression"
-      - "microtest"
-  full:
-    <<: *micro_devruntime_quick
-    labels:
-      - "rccl"
-      - "full"
-      - "full_regression"
-      - "microtest"
-
-execution_settings:
-  timeout_multiplier: 1
-  category_timeouts:
-    quick: 600
-    standard: 600
-    comprehensive: 600
-    full: 600
